Indra Card Replacement Deployed and Tested

+++ database/migrations/2023_09_22_233725_create_cl_card_rep.php
@@ -26,7 +26,7 @@ return new class extends Migration {
             $table->text('tid');
             $table->text('pax_first_name');
             $table->text('pax_last_name')->nullable();
-            $table->integer('pax_mobile');
+            $table->bigInteger('pax_mobile');
             $table->timestamp('created_at')->useCurrent();
         });
